Refactor profile-popup into separate containers for profile and curator support popups
