+++ frontend/src/features/user-management/ui/components/curator-support-button-block/index.tsx
