+++ frontend/src/features/user-management/ui/containers/curator-support-popup-container/logic.ts
@@ -1,0 +1,13 @@
+import { UIEventHandler } from '../../../../../main-ui/classes/logic'
+import ProfilePopupLogic from '../profile-popup-container/logic'
+import {
+    ProfilePopupContainerDependencies,
+    ProfilePopupContainerState,
+    ProfilePopupContainerEvent,
+} from '../profile-popup-container/types'
+
+export type CuratorSupportPopupContainerDependencies = ProfilePopupContainerDependencies
+export type CuratorSupportPopupContainerState = ProfilePopupContainerState
+export type CuratorSupportPopupContainerEvent = ProfilePopupContainerEvent
+
+export default class CuratorSupportPopupContainerLogic extends ProfilePopupLogic {}
