+++ frontend/src/features/user-management/ui/containers/curator-support-popup-container/types.ts
@@ -1,0 +1,9 @@
+import {
+    ProfilePopupDependencies,
+    ProfilePopupState,
+    ProfilePopupEvent,
+} from '../profile-popup/types'
+
+export type CuratorSupportPopupDependencies = ProfilePopupDependencies
+export type CuratorSupportPopupState = ProfilePopupState
+export type CuratorSupportPopupEvent = ProfilePopupEvent
